Refactor app commands (#8446)
The app implementation had grown to be a giant wall of text, which was difficult to read.

This change keeps each top-level command declaration in place, which maintain top-level readability of the cli. It moves each sub-command implementation to be co-located with the *Client funcs that are utilized by the sub command.

In most case is is a simple move of the sub command declaration into an existing file. In a few cases, a new file is introduced to co-locate the sub commands and client funcs that are called therein.
